% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funguildr.R
\docType{data}
\name{funguild_testdb}
\alias{funguild_testdb}
\alias{nemaguild_testdb}
\title{Short Extracts of the FUNGuild and NEMAGuild Databases}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 10 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22 rows and 10 columns.
}
\source{
FUNGuild and NEMAGuild databases, \url{http://www.stbates.org}.
}
\usage{
funguild_testdb

nemaguild_testdb
}
\description{
These are used in the examples for \code{\link[=funguild_assign]{funguild_assign()}}/\code{\link[=nemaguild_assign]{nemaguild_assign()}};
They are incomplete and should not be used for any purpose beyond testing.
}
\references{
Nguyen NH, Song Z, Bates ST, Branco S, Tedersoo L, Menke J,
Schilling JS, Kennedy PG. 2016. \emph{FUNGuild: An open annotation tool for
parsing fungal community datasets by ecological guild}. Fungal Ecology
20:241–248.
}
\keyword{datasets}
