% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funguildr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{\%<>\%}
\alias{na.omit}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%<>\%}}, \code{\link[magrittr]{\%>\%}}}

  \item{stats}{\code{\link[stats]{na.omit}}}
}}

