% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funguildr.R
\docType{data}
\name{sample_fungi}
\alias{sample_fungi}
\alias{sample_nema}
\title{Short Tables of Organisms, Used for Testing FUNGuild/NEMAGuild}
\format{
A \code{\link[tibble]{tibble}} with columns \code{Common.Name}, \code{Species}, and \code{Taxonomy}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 3 columns.
}
\source{
Taxonomy from \href{https://www.gbif.org/}{Global Biodiversity Inventory Facility} via \code{rgbif::\link[rgbif]{name_backbone}}.
}
\usage{
sample_fungi

sample_nema
}
\description{
Each of these tables contains the common name (if any), scientific name, and
taxonomic classification of a few organisms from their respective groups
(e.g., Fungi or Nematoda). They provide an example of proper formatting for
input to \code{\link{funguild_assign}} and \code{\link{nemaguild_assign}}.
}
\keyword{datasets}
