% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funguildr.R
\name{funguild_assign}
\alias{funguild_assign}
\alias{nemaguild_assign}
\title{Assign Guilds to Organisms Based on Taxonomic Classification}
\usage{
funguild_assign(otu_table, db = get_funguild_db())

nemaguild_assign(otu_table, db = get_nemaguild_db())
}
\arguments{
\item{otu_table}{A \link[base]{data.frame} with a \code{character} column named \code{Taxonomy}.
Each entry should be a comma-, colon- or semicolon-delimited classification
of an organism, from higher-ranked to lower-ranked taxa.
For example: \code{"Fungi;Dikarya;Basidiomycota;Agaricomycotina;Agaricomycetes;Agaricaceae;Agaricus;Agaricus bisporus"} (for the common cultivated mushroom).}

\item{db}{A \link[base]{data.frame} representing the FUNGuild or NEMAGuild database,
as returned by \link{get_funguild_db} or \link{get_nemaguild_db}.
If not supplied, the default database will be downloaded.}
}
\value{
A \link[tibble]{tibble} containing the original \code{otu_table},
plus relevant columns of information from the FUNGuild or NEMAGuild database.
}
\description{
These functions have identical behavior if supplied with a database; however
they download the database corresponding to their name by default.
}
\details{
Taxa present in the database are matched to the taxa present in the supplied
\code{otu_table} by exact name.  In the case of multiple matches, the lowest
(most specific) rank is chosen.  No attempt is made to check or correct the
classification in \code{otu_table}.
}
\examples{
test_table <- tibble::tribble(~Common.Name, ~Taxonomy,
                      "Button mushroom", "Agaricomycetes;Agaricales;Agaricaceae;Agaricus;Agaricus bisporus",
                      "Chanterelle", "Agaricomycetes;Cantharellales;Cantharellaceae;Cantharellus;Cantharellus cibarius",
                      "Death Cap", "Agaricomycetes;Agaricales;Amanitaceae;Amanita;Amanita phalloides",
                      "Beer Yeast", "Saccharomycetes; Saccharomycetales;Saccharomycetaceae;Saccharomyces; Saccharomyces cerevesiae")
funguild_assign(test_table)
}
