% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funguildr.R
\name{funguild_query}
\alias{funguild_query}
\title{Return entries in the FUNGuild database which match search terms}
\usage{
funguild_query(
  text,
  field = c("taxon", "guid", "mbNumber", "trophicMode", "guild", "growthForm", "trait"),
  db = "https://mycoportal.org/funguild/services/api/db_return.php"
)
}
\arguments{
\item{text}{A \code{character} string giving the text to search for. The "\verb{\%}" and
"\code{*}" characters can be used as wildcards.}

\item{field}{A \code{character} string giving the field of the database to search
for the query \code{text}. Should be one of \code{c("taxon", "guid", "mbNumber", "trophicMode", "guild", "growthForm", "trait")}.}

\item{db}{Either a \code{character} string giving the base URL of the FUNGuild
web API, or a \code{data.frame} containing a cached copy of the database,
as returned by \code{\link[=get_funguild_db]{get_funguild_db()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble}} containing all the entried from the database
which match the query.
}
\description{
Return entries in the FUNGuild database which match search terms
}
\examples{

funguild_query("Symbiotroph", "trophicMode", funguild_testdb)
}
\references{
Nguyen NH, Song Z, Bates ST, Branco S, Tedersoo L, Menke J,
Schilling JS, Kennedy PG. 2016. \emph{FUNGuild: An open annotation tool for
parsing fungal community datasets by ecological guild}. Fungal Ecology
20:241–248.
}
